requirements.txt
pyrogram==2.0.106
tgcrypto
yt-dlp
mutagen
pillow
